% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{cov}
\alias{cov}
\title{cov}
\format{An object of class \code{data.frame} with 456 rows and 4 columns.}
\usage{
cov
}
\description{
Example covariate data frame, available via data(exampleData).
}
\details{
A data frame containing the covariate information. The first two columns are required to be named “ped” and “person”, which are used to match subjects to their data in the pedigree data frame.
}
\seealso{
\code{\link{Ped}}, \code{\link{geno}}, \code{\link{exampleData}}, \code{\link{snpPos}}
}
\keyword{datasets}
