% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonFunctions.R
\name{dRule}
\alias{dRule}
\title{dRule}
\usage{
dRule(geno.only)
}
\arguments{
\item{geno.only}{The input matrix of SNP genotypes, coded 0, 1, 2.}
}
\description{
This function applies the dynamic rule to determine the number of basis functions to use
}
