% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Example data for the PedGLMM package}
\format{Four data frames: Ped, geno, cov, snpPos}
\usage{
data(exampleData)
}
\value{
None
}
\description{
Example data for the PedGLMM package
}
\examples{
data(exampleData)
dim(Ped)
head(Ped)
dim(geno)
head(geno[,1:10])
dim(cov)
head(cov)
dim(snpPos)
head(snpPos)
}
\seealso{
\code{\link{Ped}}, \code{\link{geno}}, \code{\link{cov}}, \code{\link{snpPos}}
}
