% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{geno}
\alias{geno}
\title{geno}
\format{An object of class \code{data.frame} with 456 rows and 313 columns.}
\usage{
geno
}
\description{
Example genotype data frame, available via data(exampleData).
}
\details{
A data frame containing the genotype information.  This is a matrix with genotypes for subjects (rows) at each variant position (columns). The first two columns are required to be named “ped” and “person”, which are used to match subjects to their data in the pedigree data.frame. The genotypes are coded as 0, 1, 2 for autosomal markers (typically a count of the number of the minor alleles).
}
\seealso{
\code{\link{Ped}}, \code{\link{exampleData}}, \code{\link{cov}}, \code{\link{snpPos}}
}
\keyword{datasets}
