% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonFunctions.R
\name{M_GAO}
\alias{M_GAO}
\title{M_GAO}
\source{
http://simplem.sourceforge.net/
}
\usage{
M_GAO(SNP_mx)
}
\arguments{
\item{SNP_mx}{A matrix of polymorphic SNPs (coded 0, 1, 2) with SNPs in rows and individuals in columns.}
}
\description{
Compute the effective number of independent SNPs in a region.
}
\references{
Gao X, Starmer J and Martin ER (2008) A Multiple Testing Correction Method for Genetic Association Studies Using Correlated Single Nucleotide Polymorphisms. Genetic Epidemiology 32:361-369
}
