% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega2PedGFLMM.R
\name{Mega2PedGFLMM}
\alias{Mega2PedGFLMM}
\title{Execute the PedGFLMM_beta_smooth_only function on a transcript ranges}
\usage{
Mega2PedGFLMM(gs = 1:100, genes = NULL, envir = ENV)
}
\arguments{
\item{gs}{a subrange of the default transcript ranges over which to calculate the \emph{DOPedGFLMM} function.}

\item{genes}{a list of genes over which to calculate the \emph{DOPedGFLMM} function.
The value, "*", means use all the transcripts in the selected Bioconductor database.
If genes is NULL, the gs range of the internal \emph{refRanges} will be used.}

\item{envir}{'environment' containing SQLite database and other globals}
}
\value{
None
 the data frame with the PedGFLMM_beta_smooth_only results is stored in the environment and named \emph{PedGFLMM_results},
 viz. envir$PedGFLMM_results
}
\description{
This example function illustates how to use functions from the \code{Mega2R} R package to
 iterate over defined gene ranges, computing the \code{PedGFLMM_beta_smooth_only} statistics
 for each gene that contains more than two polymorphic markers.

Execute the PedGFLMM_beta_smooth_only function on the first \emph{gs} default gene transcript ranges (gs = 1:100).
 Update the \emph{envir$PedGFLMM_results} data frame with the results.
}
\examples{
db = system.file("exdata", "seqsimmGFLMM.db", package="PedGFLMM")
ENV = init_PedGFLMM(db)
ENV$verbose = TRUE
Mega2PedGFLMM(gs = 50:60)

}
\seealso{
\code{\link{init_PedGFLMM}}
}
