% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{Ped}
\alias{Ped}
\title{Ped}
\format{An object of class \code{data.frame} with 456 rows and 7 columns.}
\usage{
Ped
}
\description{
Example pedigree data frame, available via data(exampleData)
}
\details{
A data frame containing the pedigree information with the following columns:
\describe{
  \item{ID}{Person ID}
  \item{ped}{pedigree ID,character or numeric allowed.}
  \item{person}{person ID, a unique ID within each pedigree, numeric or character allowed.}
  \item{father}{father ID, NA if no father.}
  \item{mother}{mother ID,  NA if no mother.}
  \item{sex}{sex, coded as 1 for male, 2 for female.}
  \item{trait}{trait phenotype, either case-control status coded as 1 for affected and 0 for unaffected. Subjects with missing (NA) will be removed from the analysis.}
}
}
\seealso{
\code{\link{exampleData}}, \code{\link{geno}}, \code{\link{cov}}, \code{\link{snpPos}}
}
\keyword{datasets}
