% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedGFLMM_beta_smooth_only.R
\docType{package}
\name{PedGFLMM}
\alias{PedGFLMM}
\title{PedGFLMM package}
\description{
This package implements family-based additive generalized linear mixed
models (GLMM) and generalized functional linear mixed models (GFLMM) for
gene-based association testing of dichotomous traits (Jiang et al, 2020).
}
\author{
Yingda Jiang, Chi-Yang Chiu, Daniel E. Weeks, Ruzong Fan
}
