% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{snpPos}
\alias{snpPos}
\title{snpPos}
\format{An object of class \code{data.frame} with 311 rows and 3 columns.}
\usage{
snpPos
}
\description{
Example marker position data frame, available via data(exampleData).
}
\details{
This data frame provides marker positions for each SNP. The first column, chr, contains
the chromosome number, the second column, snp, contains the SNP name, and the third
column, pos, contains the position of the SNP in base pairs.
}
\seealso{
\code{\link{Ped}}, \code{\link{geno}}, \code{\link{cov}}, \code{\link{exampleData}}
}
\keyword{datasets}
